+++ lib/glimmer/dsl/swt/widget_listener_expression.rb
@@ -8,17 +8,17 @@ module Glimmer
         include_package 'org.eclipse.swt.widgets'
   
         def can_interpret?(parent, keyword, *args, &block)
-          Glimmer.logger&.debug "keyword starts with on_: #{keyword.start_with?('on_')}"
+          Glimmer::Config.logger&.debug "keyword starts with on_: #{keyword.start_with?('on_')}"
           return false unless keyword.start_with?('on_')
           widget_or_display_parentage = widget?(parent) || parent.is_a?(Glimmer::SWT::DisplayProxy)
-          Glimmer.logger&.debug "parent is a widget or display: #{widget_or_display_parentage}"
+          Glimmer::Config.logger&.debug "parent is a widget or display: #{widget_or_display_parentage}"
           return false unless widget_or_display_parentage
-          Glimmer.logger&.debug "block exists?: #{!block.nil?}"
+          Glimmer::Config.logger&.debug "block exists?: #{!block.nil?}"
           raise Glimmer::Error, "Listener is missing block for keyword: #{keyword}" unless block_given?
-          Glimmer.logger&.debug "args are empty?: #{args.empty?}"
+          Glimmer::Config.logger&.debug "args are empty?: #{args.empty?}"
           raise Glimmer::Error, "Invalid listener arguments for keyword: #{keyword}(#{args})" unless args.empty?
           result = parent.can_handle_observation_request?(keyword)
-          Glimmer.logger&.debug "can add listener? #{result}"
+          Glimmer::Config.logger&.debug "can add listener? #{result}"
           raise Glimmer::Error, "Invalid listener keyword: #{keyword}" unless result
           true
         end